+++ core/src/test/scala/io/github/scalaquest/core/model/behaviorBased/SimpleRoomTest.scala
@@ -3,11 +3,11 @@ package io.github.scalaquest.core.model.behaviorBased
 import org.scalatest.wordspec.AnyWordSpec
 import io.github.scalaquest.core.TestsUtils.targetRoom
 import io.github.scalaquest.core.model.Direction
-import io.github.scalaquest.core.model.behaviorBased.impl.SimpleModel.roomBuilder
+import io.github.scalaquest.core.model.behaviorBased.impl.SimpleModel.Room
 
 class SimpleRoomTest extends AnyWordSpec {
   "A Room" when {
-    val room = roomBuilder("room", Map(Direction.North -> targetRoom.ref), Set())
+    val room = Room("room", Map(Direction.North -> targetRoom.ref), Set())
 
     "someone accesses a neighbor" should {
       "return a Room if the direction has an associated Room" in {
