+++ core/src/test/scala/io/github/scalaquest/core/model/behaviorBased/common/itemBehaviors/impl/SimpleEatableTest.scala
@@ -6,7 +6,6 @@ import io.github.scalaquest.core.model.{ItemDescription, ItemRef}
 import io.github.scalaquest.core.model.behaviorBased.impl.SimpleModel.{
   SimpleEatable,
   SimpleGenericItem,
-  StateUtils,
   itemsLens,
   matchRoomsLens,
   playerBagLens,
